separated left and right legs, added robot driver
This fixed the earlier issue found by the legs being mirrored. Now,
during instantiation, you simply pass in LEFT or RIGHT based on which
side the leg is on. We still need to find the correct pwm values for
these. Also added a simple robot driver which controls all 6 legs at
once.