Added table support and cleaned up design
Changed the structure of the driver to be more table-oriented as well as
more easily safety-checked. Also added functionality for rotating where
our robot's "front" is so that we can just implement a single "walk
forward" and just rotate where "forward" is. still need to actually
write it, but it will be easy.