Adds code to take a move set in and extra debug
This adds in code to take move sets and then execute the whole move set.
It also allows for re-defining the "front" of the robot. All code has
safety checking in it which is dependant on the tables found in
hex_walker_data.py. Also, renamed the "robot" class to the "hex walker"
class since "robot" will most likely be used to describe the hex walker
+ torso. also added in some basic debug stuff so that it is possible to
test and debug on a computer with no robot available.