Cleans up code by combining dictionaries
adds a few fields to the hex walker class so that we can just maintain
a single dictionary of hex_walker_position_num : hex_walker_position
rather than a separate dictionary for each attribute. Will hopefully
make our code easier to add to as well as reduce the chances of bugs due
to mispelling or mismatched dictionary values.