add torso controls and finished hex controls
added all of the torso commands and finished all of the hexapod
commands. they now are a single node but have separate callbacks for
each. torso_motion subscribes to torso_command and publishes to
torso_command_finished while the hexapod motion is subscribed to
motion_comand and publishes to motion_command_finished